add support wp-cli

+++ src/Setup/CliSupportTrigger.php
@@ -19,6 +19,11 @@ class CliSupportTrigger
 
     public function handle(): void
     {
+        if (!function_exists('shell_exec')) {
+            update_option(TwintConstant::CONFIG_CLI_SUPPORT_OPTION, 'No');
+            return;
+        }
+
         try {
             update_option(TwintConstant::CONFIG_CLI_SUPPORT_OPTION, 'No');
 
@@ -27,9 +32,7 @@ class CliSupportTrigger
             $name = escapeshellarg(CliCommand::COMMAND);
             $command = "php {$subCommand} {$name} > {$logFile} 2>&1 &";
 
-            if (function_exists('shell_exec')) {
-                shell_exec($command);
-            }
+            shell_exec($command);
         } catch (Throwable $e) {
             $this->logger->error('Cannot start PHP process: ' . $e->getMessage());
         }
